% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcrw.r
\name{fit_ssm}
\alias{fit_ssm}
\title{Correlated Random Walk Filter}
\usage{
fit_ssm(d, subset = rep(TRUE, nrow(d)), tstep = 2.4/24, nu = 5,
  gamma = 0.5, parameters = NULL, optim = c("nlminb", "optim"),
  verbose = FALSE, amf = amfCRAWL(), span = 0.3)
}
\arguments{
\item{d}{a data frame of observations}

\item{subset}{a logical vector indicating the subset of data records to be filtered}

\item{tstep}{the time step to predict to (in days)}

\item{nu}{degrees of freedom parameter}

\item{gamma}{the autocorrelation parameter used to estimate
initial parameters}

\item{parameters}{the TMB parameter list}

\item{optim}{numerical optimizer}

\item{verbose}{report progress during minimization}

\item{amf}{ARGOS error multiplication factors}

\item{span}{the span parameter for the loess fits used to estimate
initial locations}
}
\value{
a list with components
\item{\code{predicted}}{a data.frame of predicted location states}
\item{\code{fitted}}{a data.frame of fitted locations}
\item{\code{par}}{model parameter summmary}
\item{\code{data}}{the input data.frame}
\item{\code{subset}}{the input subset vector}
\item{\code{tstep}}{the prediction time step}
\item{\code{opt}}{the object returned by the optimizer}
\item{\code{tmb}}{the TMB object}
\item{\code{aic}}{the calculated Akaike Information Criterion}
}
\description{
Correlated Random Walk Filter
}
\details{
Fit a correlated random walk to filter a track and predict
locations on a regular time step.

The input track is given as a dataframe where each row is an
observed location, with columns
\describe{
\item{'date'}{observation time (as GMT POSIXct),}
\item{'lon'}{observed longitude,}
\item{'lat'}{observed latitude,}
\item{'lc'}{location class.}
}

The TMB parameter list can be specified directly with the
\code{parameters} argument. Otherwise suitable parameter values
are estimated from predicted locations estimated by fitting loess
smooths to the raw observations.

The filtering model assumes the errors in longitude and latitude
are proportional to scale factors determined by the location
class. The scale factors are specified through the \code{amf}
argument. By default the function uses the same scaling factors
for location accuracy as used in the R package crawl for ARGOS data.
}
\examples{

# Fit DCRW model for state filtering and regularization
data(ellie)
fit <- fit_ssm(ellie, tstep = 6 / 24)

## plot predicted track over observations
plot(lat ~ lon, fit$data, pch = 3, col = "firebrick")
points(lat ~ lon, fit$predicted, pch = 19, cex = 0.5, col = "dodgerblue")

## plot residuals for longitude and latitude
layout(matrix(1:2, 2, 1))
lon.res <- fit$data$lon - fit$fitted$lon
lat.res <- fit$data$lat - fit$fitted$lat
plot(fit$data$date, lon.res, xlab = "")
abline(h = 0, lty = 2, col = "red")
plot(fit$data$date, lat.res, xlab = "date")
abline(h = 0, lty = 2, col = "red")


}

